Rework argparse handling
Small changes to the User input handling mechanism. Create the course group
for mutually exclusive optins and then the outer group.

This grouping is required since the script currently handles only one action per
invokation. That is, it is notpossible to add some courses and then sync. This
feature needs to be added at a later time.