Script wil noww work from outside the Campus subnet.
This commit adds functionality to allow the script to access the CMS
portal using its external IP in case the machine is not within the campus IP
range. This uses a hackish mechanism to detect the path.

The script will check if the local IP starts with 172.16. If so, it uses the
local CMS IP, else it will use the external IP. This mechanism has many flaws,
that hopefully no one will ever see.