Python 3.2 compatibility fix
This commit fixes a compatibility issue with Python 3.2

The FileNotFoundError and related family of exceptions were added
by PEP 3151 in Python 3.3. Until v3.2, we must use the archaic OSError
and determine the cause of the exception by comparing against errno, like
is done in C programs. That code has been added.