Refactoring of download_all_files
*WARNING: Backwards incompatible change!*
This commit introduces a change which renders the current version
of the script incompatible with the log file from the previous version.
Please ensure you have deleted the old logfile manually.
TODO: A check to ensure that the old logfiles are not read and instead deleted.

The logfile name is no longer deeply coded in the source. Instead
it is referenced through a global variable now.

Split download_all_files

Split the method and separate the function _download_one_file.
This method has a single self-contained job and should be separated. However,
this refactoring is not complete. The logging should, in theory be done by
download_one_file. Also, it should accept a location parameter and download to
that location automatically.